PLANNER-2002: Add Substitute_LambdaBeanPropertyMemberAccessor (#2378)
The missing context for why Substitute_MemberAccessorFactory was
created: LambdaBeanPropertyMemberAccessor generate classes at runtime
(via LambdaMetafactory), which is not supported by GraalVM. Thus, if
a getter method is annotated for member access, a
LambdaBeanPropertyMemberAccessor is created, which will try to create
a new class, and GraalVM will throw an unsupported feature exception
at runtime when run in native mode.

The Substitute_MemberAccessorFactory "fixed" it by removing the code
path that leads to LambdaBeanPropertyMemberAccessor. However, that
is more work to maintain and requires knowledge of why
LambdaBeanPropertyMemberAccessor should not be used. What this commit
does instead, is substitute LambdaBeanPropertyMemberAccessor, keeping
the original MemberAccessorFactory implementation. The methods that
create the new classes now return a delegator class, which use
reflection to invoke the method. This allows
LambdaBeanPropertyMemberAccessor to be safely used in
MemberAccessorFactory and elsewhere in the code.