PLANNER-928: Add KOptListMoveSelector for list variables (#2538)
The K-Opt Move and Selector is based on the code in the paper 
"An Effective Implementation of K-opt Moves for the Lin-Kernighan TSP Heuristic" 
by Keld Helsgaun (http://webhotel4.ruc.dk/~keld/research/LKH/KoptReport.pdf). 

K-Opts in the paper are performed as a series of 2-Opts that are equivalent to the K-Opt. 
The 2-Opts are found by finding the signed permutation corresponding to the K-Opt, 
then finding the minimal amount of signed-reversals to convert that to the identity permutation 
(see 5.4 'Execution of a feasible move' in the paper for more details). 
Each one of those signed-reversals corresponds to a 2-opt move, which reverses a section of the route.