Switch to using @ConfigMapping interfaces (#3177)
Legacy @ConfigRoot classes support will be dropped soon from Quarkus and
Quarkus 3.15+ already supports @ConfigMapping interfaces so we can move
this project to the new infrastructure.

One thing that is important is that you are currently using runtime
config at static init - which means that it's build time for native
images and this is not correct - and disallowed.

I fixed it but the important consequence of this is that the solver will
be initialized at runtime, including when building native image.

This commit is still incomplete though as I end up with some class loader
issues and I didn't have the time to go further.